No. of annual dividend payments for IBEX set to 2
IBEX stocks will be dealt with as if paying dividends half-yearly (does not fit to all 35 components; better solution requires additional stock information about yearly no. of ex dates [e. g. 1, 2, 4, 12] in stocks_userlist.py to enhance accuracy). In the meantime this does job, but of course, this has an impact on IBEX dividend yield accuracy until further updates (example: yield is doubled for 1-time-a-year-payer or cut in half for those who stick to 4 payments such as US and most GB stocks).